Restructure app; create slack_overflow package
