--- config.py.example
@@ -1,3 +1,0 @@
-stackexchange = dict(
-    api_key = ''
-)
