+++ slack_overflow/util.py
@@ -1,0 +1,10 @@
+import os
+
+
+def envvar(name, default):
+    """Create callable environment variable getter
+
+    :param str name: Name of environment variable
+    :param default: Default value to return in case it isn't defined
+    """
+    return lambda: os.environ.get(name, default)
